requirements.txt
# --------------------------------------------------------------------------
# Azure App Service Python Production Environment
# Minimum Supported Python Version: 3.14.x
# 
# These packages have been audited and verified for Python 3.14.
# Do not downgrade core dependencies (FastAPI, Pydantic, SQLAlchemy)
# as they heavily leverage modern Python optimizations for production.
# --------------------------------------------------------------------------

alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.133.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
msal==1.35.1
packaging==26.0
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
redis==7.2.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.47
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
gunicorn==23.0.0
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.6
